+++ fv3gfs/util/_legacy_restart.py
@@ -139,7 +139,7 @@ def prepend_label(filename, label=None):
 def load_partial_state_from_restart_file(
     file, restart_properties: RestartProperties, only_names=None
 ):
-    ds = xr.open_dataset(file).isel(Time=0).drop("Time")
+    ds = xr.open_dataset(file).isel(Time=0).drop_vars("Time")
     state = map_keys(ds.data_vars, _get_restart_standard_names(restart_properties))
     state = _apply_restart_metadata(state, restart_properties)
     if only_names is None: