+++ tests/quantity/test_storage.py
@@ -137,3 +137,53 @@ def test_accessing_storage_does_not_break_view(
     )
     quantity.storage[origin] = -1.0
     assert quantity.data[origin] == quantity.view[tuple(0 for _ in origin)]
+
+
+# run using cupy backend even though unused, to mark this as a "gpu" test
+@pytest.mark.parametrize("backend", ["cupy"], indirect=True)
+def test_numpy_data_becomes_cupy_with_gpu_backend(
+    data, origin, extent, dims, units, backend
+):
+    cpu_data = np.zeros(data.shape)
+    quantity = fv3gfs.util.Quantity(
+        cpu_data,
+        origin=origin,
+        extent=extent,
+        dims=dims,
+        units=units,
+        gt4py_backend="gtcuda",
+    )
+    assert isinstance(quantity.data, cupy.ndarray)
+    assert isinstance(quantity.storage, gt4py.storage.storage.GPUStorage)
+
+
+@pytest.mark.parametrize("backend", ["gt4py_numpy"], indirect=True)
+def test_cannot_use_cpu_storage_with_gpu_backend(
+    data, origin, extent, dims, units, backend
+):
+    assert isinstance(data, gt4py.storage.storage.CPUStorage)
+    with pytest.raises(TypeError):
+        fv3gfs.util.Quantity(
+            data,
+            origin=origin,
+            extent=extent,
+            dims=dims,
+            units=units,
+            gt4py_backend="gtcuda",
+        )
+
+
+@pytest.mark.parametrize("backend", ["gt4py_cupy"], indirect=True)
+def test_cannot_use_gpu_storage_with_cpu_backend(
+    data, origin, extent, dims, units, backend
+):
+    assert isinstance(data, gt4py.storage.storage.GPUStorage)
+    with pytest.raises(TypeError):
+        fv3gfs.util.Quantity(
+            data,
+            origin=origin,
+            extent=extent,
+            dims=dims,
+            units=units,
+            gt4py_backend="numpy",
+        )